Create sentuser_c.py
The LIUNX approach..."consists of two parts.  The first part is an engine written in C that runs as a separate process.
It reads image frames from the camera, writes data to files, and writes info about frame-to-frame differences to standard out.
The second part is a GUI written in Python.  This reads the info from the C process, decides when an event occurs, and produces event files.
The GUI structure was created using a GUI builder called Glade which made the .glade file...
did not try to get the Python GUI running. (chavez)